# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Live API
  version: 1.0.0
  description: |
    REST API for the Open Live broadcast production platform.

    **Authentication**: Pass an API key via `Authorization: Bearer <key>` header or `?key=<key>` query parameter.
    The following paths are exempt from authentication: `/health`, `/ready`, `/api/v1/status`, `/api/v1/server-info`.
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyQuery: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyQuery:
      type: apiKey
      in: query
      name: key
  schemas:
...
    SourceStatus:
      type: string
      enum: [active, inactive]
...
    Source:
      type: object
      required: [id, name, address, streamType, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
          description: SRT latency in milliseconds
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SourceInput:
      type: object
      required: [name, address, streamType]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
          default: inactive
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
...
    ProductionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    Error:
      type: object
      properties:
        error:
          type: string
        statusCode:
          type: integer

    # ── Guest calling (epic #208, issue #299) ─────────────────────────────────────
...
paths:
...
  /api/v1/sources:
    get:
      summary: List sources
      operationId: listSources
      tags: [Sources]
      responses:
        '200':
          description: Array of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
    post:
      summary: Create source
      operationId: createSource
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceInput'
      responses:
        '201':
          description: Created source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Hostless SRT listener port is outside this instance's leased range (see /api/v1/server-info)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >-
            Hostless SRT listener port is already held by another source or output, or no
            port is free in the range for a port-0 assignment request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: SRT port range not yet allocated from Strom — retry shortly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/productions:
    get:
      summary: List productions
      operationId: listProductions
      tags: [Productions]
      responses:
        '200':
          description: Array of productions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Production'
    post:
      summary: Create production
      operationId: createProduction
      tags: [Productions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionInput'
      responses:
        '201':
          description: Created production
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Production'
        '400':
          description: Validation error
...
  /api/v1/productions/{id}/whip/{mixerInput}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: mixerInput
        in: path
        required: true
        schema:
          type: string
        description: e.g. `video_in_1`
    post:
      summary: WHIP offer (start ingest)
      description: Initiates a WebRTC ingest session for the given mixer input slot.
      operationId: whipOffer
      tags: [WebRTC]
      requestBody:
        required: true
        content:
          application/sdp:
            schema:
              type: string
      responses:
        '201':
          description: SDP answer
          headers:
            Location:
              schema:
                type: string
          content:
            application/sdp:
              schema:
                type: string
    patch:
      summary: WHIP ICE trickle
      operationId: whipIceTrickle
      tags: [WebRTC]
      parameters:
        - name: session
          in: query
          schema:
            type: string
          description: URL-encoded session URL
      requestBody:
        required: true
        content:
          application/trickle-ice-sdpfrag:
            schema:
              type: string
      responses:
        '200':
          description: ICE candidate accepted
    delete:
      summary: WHIP session teardown
      operationId: whipDelete
      tags: [WebRTC]
      parameters:
        - name: session
          in: query
          schema:
            type: string
          description: URL-encoded session URL
      responses:
        '204':
          description: Session closed

  # ── Guest calling (epic #208, issue #299) ─────────────────────────────────────
  /api/v1/productions/{id}/guests/invites:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Create a production-scoped guest invite
      description: |
        Mints an HMAC-signed, expiring invite token. The raw token is returned
        exactly once; only its SHA-256 hash is persisted. Requires the shared
        API key. Returns 503 when `GUEST_INVITE_SECRET` is unset (feature disabled).
      operationId: createGuestInvite
      tags: [Guests]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestInviteCreate'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestInviteCreated'
        '400':
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Production not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Guest calling disabled or database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List guest invites for a production
      operationId: listGuestInvites
      tags: [Guests]
      responses:
        '200':
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GuestInvite'
        '404':
          description: Production not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Guest calling disabled or database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/productions/{id}/guests/invites/{inviteId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: inviteId
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Revoke a guest invite
      operationId: deleteGuestInvite
      tags: [Guests]
      responses:
        '204':
          description: Invite revoked
        '403':
          description: Invite belongs to a different production
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/guests/{inviteId}/join:
    parameters:
      - name: inviteId
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Guest join (token-authed)
      description: |
        Called by the guest browser via the invite link. Authenticates with the
        per-invite HMAC token in `Authorization: Bearer <token>` — NOT the shared
        API key. `whipUrl` reuses the existing WHIP proxy contract.
      operationId: guestJoin
      tags: [Guests]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Join accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestJoinResult'
        '401':
          description: Missing, invalid, or revoked invite token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Production not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Invite expired or capacity reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Guest calling disabled or database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/guests/{inviteId}/session:
    parameters:
      - name: inviteId
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Guest leaves (token-authed)
      operationId: guestLeave
      tags: [Guests]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Session ended
        '401':
          description: Missing, invalid, or revoked invite token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Guest calling disabled or database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...